Add abstraction layer for ui
