+++ gemlist.txt
@@ -1,102 +1,9 @@
-actioncable (5.1.4, 5.1.3)
-actionmailer (5.1.4, 5.1.3)
-actionpack (5.1.4, 5.1.3)
-actionview (5.1.4, 5.1.3)
-activejob (5.1.4, 5.1.3)
-activemodel (5.1.4, 5.1.3)
-activerecord (5.1.4, 5.1.3)
-activerecord-sqlserver-adapter (5.1.1)
-activesupport (5.1.4, 5.1.3)
-addressable (2.5.2)
-arel (8.0.0)
-arrayfields (4.9.2)
-bigdecimal (1.2.8)
-bindex (0.5.0)
-builder (3.2.3)
-bundler (1.15.3)
-byebug (9.1.0)
-capybara (2.17.0)
-childprocess (0.8.0)
-chronic (0.10.2)
-coderay (1.1.2)
-coffee-rails (4.2.2)
-coffee-script (2.4.1)
-coffee-script-source (1.12.2)
-concurrent-ruby (1.0.5)
-crass (1.0.3)
-did_you_mean (1.0.0)
-erubi (1.7.0, 1.6.1)
-execjs (2.7.0)
-fattr (2.3.0)
-ffi (1.9.18 x86-mingw32)
-git (1.3.0)
-globalid (0.4.1, 0.4.0)
+for runtime
+
 glu (8.3.0 x86-mingw32)
-hoe (3.16.2)
-i18n (0.9.3, 0.8.6)
-io-console (0.4.5)
-jbuilder (2.7.0)
-jquery-rails (4.3.1)
-jruby-jars (9.1.15.0)
-json (1.8.3)
-loofah (2.1.1, 2.0.3)
-mail (2.7.0, 2.6.6)
-main (5.3.2)
-map (6.6.0)
-method_source (0.9.0, 0.8.2)
-mime-types (3.1)
-mime-types-data (3.2016.0521)
-mini_mime (1.0.0)
-mini_portile2 (2.3.0)
-minitest (5.11.3, 5.11.1, 5.8.5)
-mkrf (0.2.3)
-multi_json (1.13.1)
-net-telnet (0.1.1)
-nio4r (2.2.0, 2.1.0)
-nokogiri (1.8.1 x86-mingw32)
-ocra (1.3.10)
 opengl (0.10.0 x86-mingw32)
-pg (0.21.0 x86-mingw32)
-power_assert (0.2.6)
-pry (0.11.3)
-psych (2.1.0)
-public_suffix (3.0.1)
-puma (3.11.2)
-rack (2.0.3)
-rack-test (0.8.2, 0.6.3)
-rails (5.1.4, 5.1.3)
-rails-dom-testing (2.0.3)
-rails-html-sanitizer (1.0.3)
-railties (5.1.4, 5.1.3)
-rake (12.3.0, 12.1.0, 10.4.2)
-rake-compiler (1.0.4)
-rake-compiler-dock (0.6.2)
-rb-fsevent (0.10.2)
-rb-inotify (0.9.10)
-rb-readline (0.5.5)
-rdoc (4.2.1)
-ruboto (1.6.1)
-rubyzip (1.2.1)
-sass (3.5.5, 3.5.1)
-sass-listen (4.0.0)
-sass-rails (5.0.7, 5.0.6)
-selenium-webdriver (3.8.0)
-sfml-audio-fiddle (0.1.0)
-sprockets (3.7.1)
-sprockets-rails (3.2.1)
-sqlite3 (1.3.13 x86-mingw32)
-sqlite3-ruby (1.3.3)
-test-unit (3.1.5)
-thor (0.20.0)
-thread_safe (0.3.6)
-tilt (2.0.8)
-tiny_tds (2.0.0 x86-mingw32)
-turbolinks (5.1.0)
-turbolinks-source (5.1.0)
-tzinfo (1.2.4, 1.2.3)
-tzinfo-data (1.2017.3)
-uglifier (4.1.4)
-web-console (3.5.1)
-websocket-driver (0.6.5)
-websocket-extensions (0.1.3, 0.1.2)
-xpath (3.0.0)
+
+for executable file
+
+git(1.3.0)
+ocra(1.3.10)
